pyproject.toml
[project]
name = "de-gcp-tlc-pipeline"
version = "0.1.0"
description = "End-to-end Data Engineering pipeline on GCP (GCS + BigQuery + Airflow/Composer) using NYC TLC data."
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2.0",
  "pyarrow>=16.0.0",
  "requests>=2.32.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "google-cloud-storage>=2.17.0",
  "google-cloud-bigquery>=3.25.0",
  "db-dtypes>=1.2.0",
  "rich>=13.7.1",
  "pyyaml>=6.0"
]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "B", "UP", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
